+++ HISTORY.md
@@ -90,6 +90,9 @@
     * Dutch, thanks, Hans!
     * Brazilian Portuguese, thanks, Artur!
 
+### 2023-10-30
+* gui: added spriteNamed() helper
+
 ### 2023-10-29
 * shrunk some new built-in costumes, to make them load faster